Fixed the choppy input bug, added a 2D camera and dynamic 2D window scaling.

+++ src/main/java/dev/hilligans/engine/client/input/Input.java
@@ -48,9 +48,9 @@ public class Input implements IRegistryElement {
         return false;
     }
 
-    public void press(RenderWindow renderWindow, float strength) {}
+    public void press(RenderWindow renderWindow, double strength) {}
 
-    public void repeat(RenderWindow renderWindow, float strength) {}
+    public void repeat(RenderWindow renderWindow, double strength) {}
 
     public void bind(String key) {
         this.boundKey.add(key);
