+++ src/main/java/dev/hilligans/engine/client/input/RepeatingInput.java
@@ -6,23 +6,23 @@ import java.util.function.BiConsumer;
 
 public class RepeatingInput extends Input {
 
-    BiConsumer<RenderWindow,Float> consumer;
+    BiConsumer<RenderWindow,Double> consumer;
 
     public RepeatingInput(String name, String defaultBind) {
         super(name, defaultBind);
     }
 
-    public RepeatingInput(String name, String defaultBind, BiConsumer<RenderWindow,Float> consumer) {
+    public RepeatingInput(String name, String defaultBind, BiConsumer<RenderWindow,Double> consumer) {
         super(name, defaultBind, true);
         this.consumer = consumer;
     }
 
-    public void repeat(RenderWindow renderWindow, float strength) {
+    public void repeat(RenderWindow renderWindow, double strength) {
         consumer.accept(renderWindow,strength);
     }
 
     @Override
-    public void press(RenderWindow renderWindow, float strength) {
+    public void press(RenderWindow renderWindow, double strength) {
         consumer.accept(renderWindow,strength);
     }
 }
